Add "Flow of the day" section to the Front Page
A new section titled "Flow of the day" has been added to the Front Page. Modifications were made to the stylesheet and the front-page template, and updates were made to post type in the section "News and Notes". The "Flow of the day" section includes new formatting and provides posts about the flow of the day.