Add hover effect to 'moretag' class in global SCSS and CSS files
In the c5connections WordPress theme, a hover effect has been added to the 'moretag' class. This change has been made in both the style.css and _global.scss files, which are used to style the site. This update will make the 'moretag' change color when hovered over, improving the user interface and experience of the site.